requirements.txt
torch>=2.0.0
transformers>=4.40.0
huggingface_hub>=0.21.0
Pillow>=9.0.0
gradio>=4.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.9
accelerate>=0.27.0
